Update the TODO file.
Add some flags when we are being bitchy.

Some misc clean ups, converting things to size_t.

Signed->unsigned clean ups.

Kill many of the random unions scattered through the engine which just
happen to almost exactly match float_int_t.

Removed some fill variables in the model loader, as I'm no longer
searching for smashing. (Why did I commit those?)

New types, Sint and Uint, signed and unsigned int.

<qualifier> <type> instead of <type> <qualifier> (IE, inline void
instead of void inline)
(This is due to C standards.)

Move things around in gl_alias.c, optimize a little more.

Clean up the float->ub conversion with mod function slightly.

Optimize out some unneeded bits in the lightmap code.

Remove an unused menu option or two.

Kill some of the nastier bits of the NQ network code.

Rewrite the time tracking code to deal with reality better.
(Kills twibugs_qwsv_time_jump)

Don't draw spectators from the QW scoreboard in the minihud.

Kill random unused/unneeded cruft.

Clean up Q_sqrt.

Kill unused math funcs that just mirror the #defines.

glTexParameterf -> glTexParameteri

+++ src/base/cvar.c
@@ -332,7 +332,7 @@ int
 Cvar_CompleteCountPossible (char *partial)
 {
 	cvar_list_t	*v;
-	int			len;
+	size_t		len;
 	int			h;
 	
 	h = 0;
@@ -362,7 +362,7 @@ char	**
 Cvar_CompleteBuildList (char *partial)
 {
 	cvar_list_t	*v;
-	int			len = 0;
+	size_t		len = 0;
 	int			bpos = 0;
 	int			sizeofbuf = (Cvar_CompleteCountPossible (partial) + 1) * sizeof (char *);
 	char		**buf;
@@ -383,7 +383,7 @@ char *
 Cvar_TabComplete (const char *partial)
 {
 	cvar_list_t	   *v;
-	int				len;
+	size_t			len;
 
 	len = strlen (partial);
 	if (!len)
